units: Call /sbin/agetty instead of /bin/login
The versions of /bin/login usually present in Linux distros are actually
security tools; they do not really fit in too well as the login prompt.
For instance, the one used in Debian (from shadow-utils), while
prompting the user to input a username, terminates if the return key was
not pressed for 2 minutes, and does not process /etc/issue at all.

agetty itself is a bit overkill, and it would be best for util-linux to
provide a stripped-down version for VTs/pseudoterminals which does not
care about baudrates and other attributes of hardware terminals and
serial connections, but agetty works, so we use it.

The "-- -" at the end of the agetty command lines instead of a tty name
means "use the pts passed on standard input".

Following the principle of least surprise, we more or less match the
agetty invocation in container-getty@.service bundled with systemd.

Link: https://github.com/systemd/systemd/blob/219fa78b5fa2ec9c13dd87419366116e3197693f/units/container-getty%40.service.in