Fix running with simpledrm
The libdrm drmAvailable() function is broken, and only checks if
GPU0 is present. Systems that uses simpledrm at boot, have their
main GPU registered as GPU1, which means drm renderer is disabled,
and a black screen is displayed.
The drmAvailable() function is now deprecated, and shouldn't be used
anymore.
So remove this check, and simplify the code.

Signed-off-by: Jocelyn Falempe <jfalempe@redhat.com>